Add pages for conultant and officer obligations
