+++ app/src/pages/Obligations/index.tsx
@@ -1,43 +1,15 @@
-import React from 'react';
-import { IonContent, IonItemOption, IonIcon } from '@ionic/react';
-import getObligations from 'mocks/api/getObligations';
+import React, { ReactNode } from 'react';
+import { IonContent } from '@ionic/react';
 import Page from 'layouts/Page';
-import ObligationsList from 'components/ObligationsList';
-import { archive } from 'ionicons/icons';
 
-const Obligations = () => {
-  const obligations = getObligations();
+interface Props {
+  children: ReactNode;
+}
+
+const Obligations = ({ children }: Props) => {
   return (
     <Page title="Obligations">
-      <IonContent>
-        <ObligationsList
-          obligations={obligations}
-          startOption={
-            <IonItemOption color="danger" onClick={() => {}}>
-              <IonIcon slot="start" icon={archive} size="large" />
-            </IonItemOption>
-          }
-          endOption={
-            <IonItemOption color="success" onClick={() => {}}>
-              Complete
-            </IonItemOption>
-          }
-        />
-        <ObligationsList
-          title="Obligations pending my approval"
-          obligations={obligations}
-          startOption={
-            <IonItemOption color="danger" onClick={() => {}}>
-              Reject
-            </IonItemOption>
-          }
-          endOption={
-            <IonItemOption color="success" onClick={() => {}}>
-              Approve
-            </IonItemOption>
-          }
-        />
-      </IonContent>
+      <IonContent>{children}</IonContent>
     </Page>
   );
 };
